clean up, split stats into files
