+++ burl/src/reporting/mod.rs
@@ -1,5 +1,8 @@
 mod plots;
 mod report;
 mod stats;
+mod stats_collection;
 
 pub use report::ReportSummary;
+pub use stats::{normal_qq, percentile, standard_deviation, sum};
+pub use stats_collection::{StatsSummary, ThreadStats};
