+++ examples/test_from_toml.sh
@@ -1,2 +1,3 @@
+#!/bin/bash
 # cargo r --release -- -f "./examples/actix/specs.toml" from-toml 
 cargo r --release -- -f "./examples/fastapi/specs.toml" from-toml
